rename project to Mercury
Though the broker uses Mojolicious for an implementation, it's not
required that anyone use Mojolicious to use the broker.

Mercury is the Roman messenger god. The name Mercury also somewhat
follows the Mojolicious umbrella of names (which all start with M at
least). The Romanesque makes it fit in with Statocles/Dynamocles.
Finally, the Greek analogue for Mercury is Hermes, a Futurama character
(Mojolicious uses Futurama references extensively).

Fixes #11