add configuration file
The configuration file can set the default for the `listen` value, which
is where the broker will bind to.

This is going to be more useful when we have a Mojolicious integration,
as we will be able to use that listen value to know where to connect to
the broker.

Fixes #10