fix test failure when re-using Test::Mojo object
The Test::Mojo object, when used in WebSocket mode, stores some state
about the WebSocket to create some blocking loops, running the event
loop until an action is completed (like closing the socket).

This is a problem when we swap out the transaction in a given Test::Mojo
object: The state of the Test::Mojo object does not match the state of
the inner transaction, causing unexpected consequences.

This revealed itself when we closed a push/pull socket halfway through
the test, to ensure the round-robin was happening correctly. The next
message test would fail because the message was not yet sent.

The solution is to have a single app object, and multiple Test::Mojo
objects that all share the same app. The Test::Mojo object is used for
the duration of that socket.

Fixes #18