fix config test not finding the config file
Mojolicious::Plugin::Config has always looked for config files "relative
to the application home directory" (unless they were absolute paths).
given an app module path of `lib/Mercury.pm`, the default "application
home directory", is `lib/..`, which meant that for me developing,
`t/share/allow_origin.conf` was the correct path.

However, somehow with the advent of the Mojo::File API, what previously
worked now does not: The config file cannot be found and the test
appropriately dies with an error. This was easily fixed by absoluting
the path so that it's used as-is instead of treated as relative to the
home directory, as was my intention with this test.

That somehow the test worked before this is surprising (and incorrect),
though why I was unable to reproduce it using my standard dev
environment was obvious.

Thanks @eserte for the report!

Fixes #37